Move light declarations to effects and introduced namespace
