+++ TombEngine/Objects/Effects/effect_objects.cpp
@@ -3,9 +3,13 @@
 
 #include "Objects/Effects/flame_emitters.h"
 #include "Objects/Effects/enemy_missile.h"
+#include "Objects/Effects/light.h"
 #include "Specific/setup.h"
 
+
 using namespace TEN::Entities::Effects;
+using namespace TEN::Entities::Effects::Light;
+
 
 void InitialiseEffectsObjects()
 {
@@ -53,5 +57,42 @@ void InitialiseEffectsObjects()
 		obj->initialise = nullptr;
 		obj->collision = nullptr;
 		obj->control = ControlEnemyMissile;
+	}
+	obj = &Objects[ID_ELECTRICAL_LIGHT];
+	if (obj->loaded)
+	{
+		obj->control = ElectricalLightControl;
+		obj->drawRoutine = nullptr;
+		obj->usingDrawAnimatingItem = false;
+	}
+
+	obj = &Objects[ID_PULSE_LIGHT];
+	if (obj->loaded)
+	{
+		obj->control = PulseLightControl;
+		obj->drawRoutine = nullptr;
+		obj->usingDrawAnimatingItem = false;
+	}
+
+	obj = &Objects[ID_STROBE_LIGHT];
+	if (obj->loaded)
+	{
+		obj->control = StrobeLightControl;
+	}
+
+	obj = &Objects[ID_COLOR_LIGHT];
+	if (obj->loaded)
+	{
+		obj->control = ColorLightControl;
+		obj->drawRoutine = nullptr;
+		obj->usingDrawAnimatingItem = false;
+	}
+
+	obj = &Objects[ID_BLINKING_LIGHT];
+	if (obj->loaded)
+	{
+		obj->control = BlinkingLightControl;
+		obj->drawRoutine = nullptr;
+
 	}
 }
