+++ TombEngine/Objects/Effects/light.h
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace TEN::Entities::Effects::Light
+{
+	void PulseLightControl(short itemNumber);
+	void TriggerAlertLight(int x, int y, int z, int r, int g, int b, int angle, short room, int falloff);
+	void StrobeLightControl(short itemNumber);
+	void ColorLightControl(short itemNumber);
+	void ElectricalLightControl(short itemNumber);
+	void BlinkingLightControl(short itemNumber);
+}
