--- TombEngine/Objects/TR5/Light/tr5_light.h
@@ -1,8 +1,0 @@
-#pragma once
-
-void PulseLightControl(short itemNumber);
-void TriggerAlertLight(int x, int y, int z, int r, int g, int b, int angle, short room, int falloff);
-void StrobeLightControl(short itemNumber);
-void ColorLightControl(short itemNumber);
-void ElectricalLightControl(short itemNumber);
-void BlinkingLightControl(short itemNumber);
